[opentherm] Rename c++ files for predictable doxygen generation
