+++ esphome/components/opentherm/output/opentherm_output.h
