+++ esphome/components/opentherm/switch/opentherm_switch.h
